refactor(code-block): extract code block editing logic to custom hook
Move code block editing functionality from Messages component to useEditCodeBlock hook for better reusability and maintainability
